Despliegue .exe
Genera un único archivo .exe, ejecutable desde cualquier ubicación (puede moverse a otra carpeta). Pendiente cambiar el icono. Todos los .py movidos a carpeta application. Instrucciones generación del .exe en un txt.
