+++ app.py
@@ -1,0 +1,3 @@
+import streamlit as st
+if __name__ == '__main__':
+    st.header("Hello world")
